requirements.txt
appdirs==1.4.4
APScheduler==3.3.1
apturl==0.5.2
astroid==2.5.3
autopep8==1.5.6
bcrypt==3.1.7
bidict==0.21.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
eventlet==0.30.2
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.2
Flask-APScheduler==1.12.2
Flask-SocketIO==5.0.1
future==0.18.2
greenlet==1.0.0
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
mypy==0.812
mypy-extensions==0.4.3
netifaces==0.10.4
oauthlib==3.1.0
paramiko==2.6.0
pexpect==4.6.0
protobuf==3.6.1
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.7.4
pymacaroons==0.13.0
pymongo==3.11.3
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-engineio==4.1.0
python-socketio==5.1.0
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
socketIO-client==0.7.2
systemd-python==234
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.4.3
wadllib==1.3.3
websocket-client==0.58.0
Werkzeug==1.0.1
wrapt==1.12.1
xkit==0.0.0
